+++ Overlayer/Core/Tags/TagExtensions.cs
@@ -8,13 +8,13 @@ namespace Overlayer.Core
 {
     public static class TagExtensions
     {
-        public static float Round(this float value, float digits)
+        public static float Round(this float value, double digits)
         {
             if (digits == -1)
                 return value;
             return (float)Math.Round(value, (int)digits);
         }
-        public static float Round(this double value, float digits)
+        public static float Round(this double value, double digits)
         {
             if (digits == -1)
                 return (float)value;
